fix user controller; closes #58, closes #59
